extract "ToolbarBase" from BottomToolbar component
